refactor metadata with factory method
